+++ models/balance.go
@@ -5,12 +5,12 @@ import "time"
 type LoyaltyOrder struct {
 	ID          int64     `json:"-"`
 	Code        string    `json:"order"`         // Уникальный номер заказа
-	Deduction   int       `json:"sum,omitempty"` // Списанные бонусы в счет текущего заказа
+	Deduction   float64   `json:"sum,omitempty"` // Списанные бонусы в счет текущего заказа
 	ProcessedAt time.Time `json:"processed_at"`  // Дата время загрузки заказа
 	UserID      int64     `json:"-"`             // Связь с заказчиком
 }
 
-func NewLoyaltyOrder(code string, deduction int, userID int64) *LoyaltyOrder {
+func NewLoyaltyOrder(code string, deduction float64, userID int64) *LoyaltyOrder {
 	return &LoyaltyOrder{
 		Code:        code,
 		Deduction:   deduction,
@@ -19,6 +19,6 @@ func NewLoyaltyOrder(code string, deduction int, userID int64) *LoyaltyOrder {
 }
 
 type StatusBalance struct {
-	Current   int `json:"current"`
-	Withdrawn int `json:"withdrawn"`
+	Current   float64 `json:"current"`
+	Withdrawn float64 `json:"withdrawn"`
 }
